Significantly improve ILP performance
The LP does not have to solve for Beta. Only the ILP is now constrained by a given Beta. Before this change, the LP was producing restricted solution sets which the ILP struggled to work with and therefore required a larger Beta given a small time limit such as only 10 seconds.