Massive overhaul of LP
Setup code for the LP now accounts for a beta larger than the number of feasible guides found by the LP solver. Always start by solving a minimization without scores. When we have more feasible guides than beta, meaning that we need to reduce the set size further, pass the guides on to the ILP to solve for maximization with scores. When we have fewer feasible guides than beta, we resolve for maximization with scores in the LP, then pass on the residuals to the ILP.